split runneradmin structs into it's own package
